src/main/java/common/domain/User.java
package common.domain;

import java.util.UUID;

public class User {
    private String userID;
    private String firstName;
    private String lastName;
    private String emailAddress;
    private int age;

    public User() {
        this.userID = UUID.randomUUID().toString();
    }

    public User(String firstName, String lastName, String emailAddress, int age) {
        this.userID = UUID.randomUUID().toString();
        this.firstName = firstName;
        this.lastName = lastName;
        this.emailAddress = emailAddress;
        this.age = age;
    }

    public String getUserID() {
        return userID;
    }

    public String getFirstName() {
        return firstName;
    }

    public void setFirstName(String firstName) {
        this.firstName = firstName;
    }

    public String getLastName() {
        return lastName;
    }

    public void setLastName(String lastName) {
        this.lastName = lastName;
    }

    public String getEmailAddress() {
        return emailAddress;
    }

    public void setEmailAddress(String emailAddress) {
        this.emailAddress = emailAddress;
    }

    public int getAge() {
        return age;
    }

    public void setAge(int age) {
        this.age = age;
    }
}
